move to_hoomdsnapshot to conversion, add method in Compound class. Delete old snapshot and gsd writers
